+++ src/main/java/ru/yandex/practicum/filmorate/utils/Utils.java
@@ -1,15 +1,15 @@
 package ru.yandex.practicum.filmorate.utils;
 
 public class Utils {
-    private static int lastUserId = 0;
-    private static int lastFilmId = 0;
+    private static long lastUserId = 0;
+    private static long lastFilmId = 0;
 
-    public static int generateFilmId() {
+    public static long generateFilmId() {
         lastUserId++;
         return lastUserId;
     }
 
-    public static int generateUserId() {
+    public static long generateUserId() {
         lastFilmId++;
         return lastFilmId;
     }
